+++ crates/project-expander/src/token_expander_error.rs
@@ -6,29 +6,59 @@ use thiserror::Error;
 pub enum TokenExpanderError {
     #[diagnostic(code(token_expander::invalid_index))]
     #[error(
-        "Token {} received an invalid type for index \"{index}\", must be a number.",
-        .token.style(Style::Symbol)
+        "Token {} in task {} received an invalid type for index \"{index}\", must be a number.",
+        .token.style(Style::Symbol),
+        .target.style(Style::Label),
     )]
-    InvalidTokenIndex { token: String, index: String },
+    InvalidTokenIndex {
+        target: String,
+        token: String,
+        index: String,
+    },
 
     #[diagnostic(code(token_expander::invalid_index_reference))]
     #[error(
-        "Token {} is referencing another token or an invalid value. Only file paths or globs can be referenced by index.",
-        .token.style(Style::Symbol)
+        "Token {} in task {} is referencing another token or an invalid value. Only file paths or globs can be referenced by index.",
+        .token.style(Style::Symbol),
+        .target.style(Style::Label),
     )]
-    InvalidTokenIndexReference { token: String },
+    InvalidTokenIndexReference { target: String, token: String },
 
     #[diagnostic(code(token_expander::invalid_scope))]
-    #[error("Token {} cannot be used within task {scope}.", .token.style(Style::Symbol))]
-    InvalidTokenScope { token: String, scope: String },
+    #[error(
+        "Token {} in task {} cannot be used within task {scope}.",
+        .token.style(Style::Symbol),
+        .target.style(Style::Label),
+    )]
+    InvalidTokenScope {
+        target: String,
+        token: String,
+        scope: String,
+    },
 
     #[diagnostic(code(token_expander::missing_in_index))]
-    #[error("Input index {index} does not exist for token {}.", .token.style(Style::Symbol))]
-    MissingInIndex { index: usize, token: String },
+    #[error(
+        "Input index {index} does not exist for token {} in task {}.",
+        .token.style(Style::Symbol),
+        .target.style(Style::Label),
+    )]
+    MissingInIndex {
+        index: usize,
+        target: String,
+        token: String,
+    },
 
     #[diagnostic(code(token_expander::missing_out_index))]
-    #[error("Output index {index} does not exist for token {}.", .token.style(Style::Symbol))]
-    MissingOutIndex { index: usize, token: String },
+    #[error(
+        "Output index {index} does not exist for token {} in task {}.",
+        .token.style(Style::Symbol),
+        .target.style(Style::Label),
+    )]
+    MissingOutIndex {
+        index: usize,
+        target: String,
+        token: String,
+    },
 
     #[diagnostic(code(token_expander::unknown_file_group))]
     #[error(